+++ data/repositories/src/main/java/es/sebas1705/repositories/repos/SettingsRepository.kt
@@ -28,6 +28,7 @@ class SettingsRepository @Inject constructor(
         settingsPreferencesDataSource.saveMusicVolume(settingsData.musicVolume)
         settingsPreferencesDataSource.saveSoundVolume(settingsData.soundVolume)
         settingsPreferencesDataSource.saveAppContrast(settingsData.appContrast)
+        settingsPreferencesDataSource.saveLanguage(settingsData.language)
     }
 
 }